* Makefile: added IMGFILE; moved some stuff around. * flmodule.c: added some missing functions; changed readonly flags of   some data members based upon FORMS documentation. * listobject.c: fixed int/long arg lint bug (bites PC compilers). * several: removed redundant print methods (repr is good enough). * posixmodule.c: added (still experimental) process group functions.

+++ Objects/listobject.c
@@ -600,7 +600,7 @@ listindex(self, args)
 	}
 	for (i = 0; i < self->ob_size; i++) {
 		if (cmpobject(self->ob_item[i], args) == 0)
-			return newintobject(i);
+			return newintobject((long)i);
 	}
 	err_setstr(ValueError, "list.index(x): x not in list");
 	return NULL;
